+++ falco_event/Cargo.toml
@@ -22,6 +22,7 @@ derive_deftly = ["dep:derive-deftly"]
 
 [dependencies]
 falco_event_derive = { path = "../falco_event_derive", version = "0.4.0" }
+falco_schema_derive = { path = "../falco_schema_derive", version = "0.4.0" }
 thiserror = "2.0.12"
 bitflags = { version = "2.4.2" }
 anyhow = "1.0.81"
