+++ falco_schema_derive/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "falco_schema_derive"
+version = "0.4.0"
+edition = "2021"
+license = "Apache-2.0"
+description = "Helper proc macros for the falco_event library"
+homepage = "https://falcosecurity.github.io/plugin-sdk-rs/falco_event_derive/"
+repository = "https://github.com/falcosecurity/plugin-sdk-rs"
+readme = "README.md"
+keywords = ["falco", "security"]
+
+[lints]
+workspace = true
+
+[lib]
+proc-macro = true
+
+[dependencies]
+quote = "1"
+proc-macro2 = "1.0"
+syn = "2.0.100"
