+++ falco_schema_derive/README.md
@@ -1,0 +1,3 @@
+# Proc macros for generating Falco schema types
+
+These are not intended to be used directly.
